% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacteria_kill.R
\name{bacteria_kill}
\alias{bacteria_kill}
\title{effectiveness of disinfectant}
\usage{
bacteria_kill(num_bacteria, resistance = FALSE)
}
\arguments{
\item{num_bacteria}{bacteria population total}

\item{resistance}{logical, whether or not bacteria population has resistance to the disinfectant, FALSE by default}
}
\value{
population still alive from the pre treatment population
}
\description{
effectiveness of disinfectant
}
\examples{
bacteria_kill(50000, TRUE)
}
