% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice_cream_status.R
\name{ice_cream_status}
\alias{ice_cream_status}
\title{ice cream status}
\usage{
ice_cream_status(ambient_temp, time)
}
\arguments{
\item{ambient_temp}{temperature outside (F)}

\item{time}{how much time has passed since u got ur ice cream (min)}
}
\value{
whether or not ur ice cream has melted
}
\description{
has ur ice cream melted yet
}
\examples{
ice_cream_status(90, 2)
}
