% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_sus_harvest_func.R
\name{max_sustainable_harvest}
\alias{max_sustainable_harvest}
\title{Max sustainable harvest}
\usage{
max_sustainable_harvest(K, r)
}
\arguments{
\item{K}{car capacity of fishery (in individuals)}

\item{r}{growth rate of fishery (1/year)}
}
\value{
max sustainable harvest (individuals/year)
}
\description{
returns max sus harvest given car capcity + growth rate
}
\examples{
max_sustainable_harvest(K = 36000, r = 0.62)

}
