% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funny.R
\name{the_funny}
\alias{the_funny}
\title{is_funny}
\usage{
the_funny(funny, laugh_level, is_funny = FALSE)
}
\arguments{
\item{funny}{string, put the funny here}

\item{laugh_level}{numeric, the level of funny}

\item{is_funny}{probably not but feel free to disagree. logical defaulted to FALSE}
}
\value{
how many funnies u awarded ur joke
}
\description{
are u funny, idk maybe
}
\examples{
the_funny(funny = "steals ur organs", laugh_level = 99999, is_funny = TRUE)
}
