% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_shrek.R
\name{theme_shrek}
\alias{theme_shrek}
\title{shrek theme}
\usage{
theme_shrek()
}
\value{
add to ggplot2 for swamp time
}
\description{
get out of my swamp
}
\details{
see documentation for...

theme(),
element_rect(),
element_text(),
element_line(),
}
\examples{

ggplot(data, aes(x, y)) + geom_point()+
 theme_shrek()
}
